fix merge issues; unpack message at node
